+++ src/main/java/com/ksolorzano/KinScription/dominio/repository/AdmReportePagoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AdmReportePagoRepository extends JpaRepository<AdmReportePago, Long> {
+public interface AdmReportePagoRepository extends JpaRepository<AdmReportePago, Integer> {
 }
