+++ src/main/java/com/ksolorzano/KinScription/persistence/entity/GradoCurso.java
@@ -11,11 +11,11 @@ public class GradoCurso {
 
     @Id
     @Column(name = "idGrado")
-    private Long idGrado;
+    private Integer idGrado;
 
     @Id
     @Column(name = "idCurso")
-    private Long idCurso;
+    private Integer idCurso;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "idGrado", insertable = false, updatable = false)
