+++ src/main/java/com/ksolorzano/KinScription/persistence/entity/ProfesorAsignacionId.java
@@ -10,8 +10,8 @@ import java.io.Serializable;
 @NoArgsConstructor
 @AllArgsConstructor
 public class ProfesorAsignacionId implements Serializable {
-    private Long idProfesor;
-    private Long idCurso;
-    private Long idGrado;
-    private Long idSeccion;
+    private Integer idProfesor;
+    private Integer idCurso;
+    private Integer idGrado;
+    private Integer idSeccion;
 }
