+++ src/main/java/com/ksolorzano/KinScription/persistence/entity/Tutor.java
@@ -19,7 +19,7 @@ public class Tutor {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_tutor")
-    private Long idTutor;
+    private Integer idTutor;
 
     @Column(name = "nombreCompleto")
     private String nombreCompleto;
